--- a/rollup.config.js
+++ b/rollup.config.js
@@ -1,0 +1,8 @@
+export default {
+  input: 'src/marked.js',
+  output: {
+    file: 'marked.min.js',
+    format: 'umd',
+    name: 'marked'
+  }
+}