--- a/rollup.config.js
+++ b/rollup.config.js
@@ -8,4 +8,4 @@ export default {
     name: 'marked'
   },
   plugins: [babel()]
-}
+};