👋 

I had this lying around and decided to finish it off recently in my fork as it was needed for trying out in some web component stuff I had setup.

Fixes #1288

Is this something you'd be interested in? Happy to keep it in my fork if not and close this PR.

---

Here's a summary of what happened in this branch:

* Sources separated into modules
* All sources are valid ESM (`src/*`)
* Moved from prototypical concepts to classes
* Introduced babel to transpile to CJS modules (`lib/*`) for when node users require this
* Introduced rollup to bundle to a browser-compatible library (`marked.min.js`) which behaves the same as the original
* Introduced `files` to package.json so `npm pack` and publish will package only bin/lib/man (CJS sources)
* Introduced a `module` entry to package.json so bundlers can resolve the ESM sources
* Replaced all unscoped vars and what not to `let`/`const`

All tests pass, build succeeds.

Have not been using it in a browser myself so it seems that would need further testing.

Whats missing:

* The browser bundle isn't minified (something like `babel-preset-minify` could achieve that)
* Performance comparison
* Build will fail on CI as `npm run build` needs adding to travis.yml

---

Again, if this is of any use to you just let me know and ill get back on it.

cc @joshbruce