@styfle 2018-10-03
The [tests](https://travis-ci.org/markedjs/marked/builds/436376656) are failing with Error: Cannot find module './lexer'