@43081j 2018-10-03
Yup this is because the travis config needs updating to `npm run build` beforehand.