@styfle 2018-10-03
@43081j You should be able to modify `.travis.yml` and add a build step