@UziTech 2018-10-03
Do we need to add [browserslist](https://babeljs.io/docs/en/babel-preset-env#browserslist-integration) to `package.json` for `babel-preset-env`?