@43081j 2018-10-03
We probably should but i wasn't too sure what the correct targets should be.

The ESM sources account for any browsers supporting ES modules.

The CJS sources should only be used by node, so our babel-preset config should probably be `node: true`.

The bundle should be consumed by non-ESM browsers so its actually a different target to `lib/` I guess (overridable via rollup.config.js if we must). We would have to find the minimum browser version we could support.