@UziTech 2018-10-03
benchmarks seem to be consistent 👍:

```
// master
marked completed in 5809ms.
marked (gfm) completed in 7043ms.
marked (pedantic) completed in 5590ms.
commonmark completed in 8877ms.
markdown-it completed in 6774ms.
markdown.js completed in 22230ms.

// this pr
marked completed in 6027ms.
marked (gfm) completed in 6981ms.
marked (pedantic) completed in 5792ms.
commonmark completed in 9057ms.
markdown-it completed in 8349ms.
markdown.js completed in 20882ms.
```