@43081j 2018-10-16
What sort of time frame would you be looking at for 1.x?

To be honest, ES modules are still a bit flaky with node which is why we're forced to have 3 builds (ESM, CJS, browser).

I think if we try wait for node to have a better implementation, we'll be here _a while_.

It may be worth us looking into using this:
https://github.com/standard-things/esm

instead of shipping two copies of the sources 🤔  but its a fairly big dependency to be shipping i guess