@styfle 2018-10-16
I looked this over again and I am becoming more reserved the more I look at it.

> ES modules are still a bit flaky with node which is why we're forced to have 3 builds (ESM, CJS, browser).

This PR is adding more complexity to a relatively small project and I'm not sure the value added here is worth it at this time.

Let's do a pros/cons list

## Pros

1. Developers get to use multiple files - maybe avoid merge conflicts?
2. Developers get to use ES6 syntax - less writing and more readable in most cases
3. (maybe) Users can use ESM sources with bundlers to maybe treeshake output

## Cons

1. The ESM source will not run in the browser as-is (due to bare specifiers)
2. The ESM source will not run in Node.js as-is (due to .mjs and experimental)
3. The bundle size is actually larger by a few KB (was 42 KB and now 47 KB)
4. We'll be publishing more sources files so removing those sources is a breaking change (probably can do this in 1.x)
5. More complexity due to more build tools requires maintenance when upgrading versions: babel, rollup, and several more plugins
6. (maybe) Some js user's don't know ES6 syntax and won't be able to contribute
7. (maybe) Build time will increase slightly due to more build targets