@43081j 2018-10-16
This is meant to introduce the ability to use the ES modules in browsers _as is_. So if that means I forgot to (and need to) add the file extension to each import, I should.

The bundle size is _of course_ larger, it isn't even minified yet (and rollup will tree-shake, which it hasn't done yet either).

The rest of @styfle's points are fairly small IMHO, tooling being introduced isn't a problem, nor is build time (it won't be noticeable most likely). Git history shouldn't be an issue either as the files should be detected as being moved.