@Martii 2018-10-16
> Developers get to use ES6 syntax - less writing and more readable in most cases

This is actually a con in our book. ES6 may be supported however it's *very subjective on readability*. Examples written in ES5 are a good base for projects that don't use ES6+ and extending it to the more quirky ES6+ syntaxes.